+++ src/sys/class/alo/traversables/smartobj.php
@@ -2,11 +2,6 @@
 
     namespace Alo\Traversables;
 
-    use IteratorAggregate;
-    use ArrayIterator;
-    use ArrayAccess;
-    use Countable;
-
     if (!defined('GEN_START')) {
         http_response_code(404);
     } else {
@@ -15,32 +10,7 @@
          * Some smarter array functions. This is a very work-in-progress class.
          * @author Art <a.molcanovas@gmail.com>
          */
-        class SmartObj implements IteratorAggregate, ArrayAccess, Countable {
-
-            /**
-             * The array we're working with
-             * @var array
-             */
-            protected $data;
-
-            /**
-             * Initialises our smart array-based object
-             * @author Art <a.molcanovas@gmail.com>
-             *
-             * @param array $initial The initial array to set
-             */
-            function __construct(array $initial = []) {
-                $this->data = $initial;
-            }
-
-            /**
-             * Returns the number of items in the data object
-             * @author Art <a.molcanovas@gmail.com>
-             * @return int
-             */
-            function count() {
-                return count($this->data);
-            }
+        class SmartObj extends ArrayObj {
 
             /**
              * Initialises our smart array-based object
@@ -55,41 +25,6 @@
             }
 
             /**
-             * Returns an array value
-             *
-             * @author Art <a.molcanovas@gmail.com>
-             *
-             * @param string $k The value's key
-             *
-             * @return mixed
-             */
-            function __get($k) {
-                return get($this->data[$k]);
-            }
-
-            /**
-             * Sets an array value
-             *
-             * @author Art <a.molcanovas@gmail.com>
-             *
-             * @param string $k The key
-             * @param mixed  $v The value
-             */
-            function __set($k, $v) {
-                $this->data[$k] = $v;
-            }
-
-            /**
-             * Returns the data set
-             *
-             * @author Art <a.molcanovas@gmail.com>
-             * @return array
-             */
-            function toArray() {
-                return $this->data;
-            }
-
-            /**
              * Removes all the duplicate values from an array
              * @author Art <a.molcanovas@gmail.com>
              *
@@ -97,6 +32,62 @@
              */
             function uniqueRecursive() {
                 $this->uniqueRecursiveInternal($this->data);
+
+                return $this;
+            }
+
+            /**
+             * Internal handler for regex deletion
+             * @author Art <a.molcanovas@gmail.com>
+             *
+             * @param string $regex       The regular expression
+             * @param bool   $recursive   Whther to recurse into child arrays
+             * @param bool   $checkingKey true if checking key regex, false if checking value regex
+             * @param bool   $inverse     If set to true the method will delete items that do NOT match the regex
+             * @param array  $currArray   Reference to the currently checked array
+             */
+            protected function deleteRegexInternal($regex, $recursive, $checkingKey, $inverse, &$currArray) {
+                foreach ($currArray as $k => &$v) {
+                    $checking = $checkingKey ? $k : $v;
+
+                    if (is_scalar($checking) &&
+                        ($inverse ? !preg_match($regex, $checking) : preg_match($regex, $checking))
+                    ) {
+                        unset($currArray[$k]);
+                    } elseif ($recursive && is_array($v)) {
+                        $this->deleteRegexInternal($regex, $recursive, $checkingKey, $inverse, $v);
+                    }
+                }
+            }
+
+            /**
+             * Deletes all elements from an array where the value matches the supplied regular expression
+             * @author Art <a.molcanovas@gmail.com>
+             *
+             * @param string $regex     The regular expression
+             * @param bool   $recursive Whether to recurse into child arrays
+             * @param bool   $inverse   If set to true the method will delete items that do NOT match the regex
+             *
+             * @return SmartObj
+             */
+            function deleteWithValueRegex($regex, $recursive = true, $inverse = false) {
+                $this->deleteRegexInternal($regex, $recursive, false, $inverse, $this->data);
+
+                return $this;
+            }
+
+            /**
+             * Deletes all elements from an array where the key matches the supplied regular expression
+             * @author Art <a.molcanovas@gmail.com>
+             *
+             * @param string $regex     The regular expression
+             * @param bool   $recursive Whether to recurse into child arrays
+             * @param bool   $inverse   If set to true the method will delete items that do NOT match the regex
+             *
+             * @return SmartObj
+             */
+            function deleteWithKeyRegex($regex, $recursive = true, $inverse = false) {
+                $this->deleteRegexInternal($regex, $recursive, true, $inverse, $this->data);
 
                 return $this;
             }
@@ -114,69 +105,6 @@
                         $this->uniqueRecursiveInternal($v);
                     }
                 }
-            }
-
-            /**
-             * Returns the array iterator for our data
-             * @author Art <a.molcanovas@gmail.com>
-             *
-             * @return ArrayIterator
-             */
-            function getIterator() {
-                return new ArrayIterator($this->data);
-            }
-
-            /**
-             * Checks whether a offset exists
-             * @author Art <a.molcanovas@gmail.com>
-             * @link   http://php.net/manual/en/arrayaccess.offsetexists.php
-             *
-             * @param mixed $offset An offset to check for.
-             *
-             * @return boolean
-             */
-            function offsetExists($offset) {
-                return array_key_exists($offset, $this->data);
-            }
-
-            /**
-             * Gets an offset
-             * @author Art <a.molcanovas@gmail.com>
-             * @link   http://php.net/manual/en/arrayaccess.offsetget.php
-             *
-             * @param mixed $offset The offset to retrieve.
-             *
-             * @return mixed
-             */
-            function offsetGet($offset) {
-                return get($this->data[$offset]);
-            }
-
-            /**
-             * Sets an offset
-             * @author Art <a.molcanovas@gmail.com>
-             * @link   http://php.net/manual/en/arrayaccess.offsetset.php
-             *
-             * @param mixed $offset The offset to assign the value to.
-             * @param mixed $value  The value to set.
-             */
-            function offsetSet($offset, $value) {
-                if ($offset === null) {
-                    $this->data[] = $value;
-                } else {
-                    $this->data[$offset] = $value;
-                }
-            }
-
-            /**
-             * Unsets an offset
-             * @author Art <a.molcanovas@gmail.com>
-             * @link   http://php.net/manual/en/arrayaccess.offsetunset.php
-             *
-             * @param mixed $offset The offset to unset.
-             */
-            function offsetUnset($offset) {
-                unset($this->data[$offset]);
             }
         }
     }